Modularize Language and Configuration
This commit includes modules for the Language and Configuration models.
