+++ soundgo_api/configuration/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ConfigurationConfig(AppConfig):
+    name = 'configuration'
