+++ soundgo_api/env_setup.sh
@@ -28,16 +28,16 @@ echo -e "\n<<<< Populating database >>>>\n"
 python3 manage.py shell -c "from records.models import Category; Category.objects.all().delete();"
 python3 manage.py shell -c "from records.models import Category; category1 = Category.objects.create(name='Leisure', maxTimeRecord=60, minDurationMap=259200); category1.save(); category2 = Category.objects.create(name='Experience', maxTimeRecord=60, minDurationMap=259200); category2.save(); category3 = Category.objects.create(name='Tourism', maxTimeRecord=60, minDurationMap=259200); category3.save();"
 
-python3 manage.py shell -c "from accounts.models import Language; Language.objects.all().delete();"
-python3 manage.py shell -c "from accounts.models import Language; language1 = Language.objects.create(name='Spanish'); language1.save(); language2 = Language.objects.create(name='French');
+python3 manage.py shell -c "from languages.models import Language; Language.objects.all().delete();"
+python3 manage.py shell -c "from languages.models import Language; language1 = Language.objects.create(name='Spanish'); language1.save(); language2 = Language.objects.create(name='French');
 language2.save(); language3 = Language.objects.create(name='Català'); language3.save(); language4 = Language.objects.create(name='Portuguese'); language4.save();
 language5 = Language.objects.create(name='Italian'); language5.save(); language6 = Language.objects.create(name='German'); language6.save(); language7 = Language.objects.create(name='Japanese');
 language7.save(); language8 = Language.objects.create(name='Chinese'); language8.save(); language9 = Language.objects.create(name='English'); language9.save();
 language10 = Language.objects.create(name='Euskera'); language10.save(); language11 = Language.objects.create(name='Galician'); language11.save();
 language12 = Language.objects.create(name='Arab'); language12.save(); language13 = Language.objects.create(name='Other'); language13.save();"
 
-python3 manage.py shell -c "from accounts.models import SoundGoConfig; SoundGoConfig.objects.all().delete();"
-python3 manage.py shell -c "from accounts.models import SoundGoConfig; sound_go_config = SoundGoConfig.objects.create(maximum_radius=2000, minimum_radius=20, time_listen_advertisement=3, minimum_reports_ban=10); sound_go_config.save();"
+python3 manage.py shell -c "from configuration.models import Configuration; Configuration.objects.all().delete();"
+python3 manage.py shell -c "from configuration.models import Configuration; config = Configuration.objects.create(maximum_radius=2000, minimum_radius=20, time_listen_advertisement=3, minimum_reports_ban=10); config.save();"
 
 echo -e "Database populated"
