+++ soundgo_api/languages/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class LanguagesConfig(AppConfig):
+    name = 'languages'
