+++ soundgo_api/soundgo_api/settings.py
@@ -41,6 +41,8 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'rest_framework',
     'accounts',
+    'configuration',
+    'languages',
     'records',
     'sites',
     'tags',
